Add CCTV7 image
CCTV-7 is a flagship free-to-air terrestrial channel of the CCTV in the People's Republic of China. It has a mixture of agricultural and military programmes and is available to both cable and terrestrial television viewers and is broadcast from the CMG Headquarters at East 3rd Ring Road off Guanghua Road in Chaoyang Metropolitan form Beijing Central Business District.